Extract pure main window menu helpers into adjacent files
